Check if download works

+++ install-nvim.sh
@@ -2,9 +2,13 @@
 RED="\e[1;41m"
 ENDCOLOR="\e[0m"
 curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux-x86_64.tar.gz
-sudo rm -rf /opt/nvim-linux-x86_64
-sudo tar -C /opt -xzf nvim-linux-x86_64.tar.gz
-rm nvim-linux-x86_64.tar.gz
-echo -e "Add nvim to your path like export ${RED}PATH=\"\$PATH:/opt/nvim-linux-x86_64/bin\"${ENDCOLOR}"
-echo -e "Install depenencies with apt  ${RED}sudo apt install luarocks fd-find ripgrep${ENDCOLOR}"
-echo -e "Install depenencies with nala ${RED}sudo nala install luarocks fd-find ripgrep${ENDCOLOR}"
+if [ -f nvim-linux-x86_64.tar.gz ]; then
+    sudo rm -rf /opt/nvim-linux-x86_64
+    sudo tar -C /opt -xzf nvim-linux-x86_64.tar.gz
+    rm nvim-linux-x86_64.tar.gz
+    echo -e "Add nvim to your path like export ${RED}PATH=\"\$PATH:/opt/nvim-linux-x86_64/bin\"${ENDCOLOR}"
+    echo -e "Install depenencies with apt  ${RED}sudo apt install luarocks fd-find ripgrep${ENDCOLOR}"
+    echo -e "Install depenencies with nala ${RED}sudo nala install luarocks fd-find ripgrep${ENDCOLOR}"
+else
+    echo "Download failed"
+fi
